backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
pydantic-settings
python-dotenv
pypdf
tiktoken
sentence-transformers
chromadb
rank-bm25
groq
pytesseract
pypdfium2
Pillow
numpy<2
packaging<26
protobuf<6
httpx
